requirements-dev.txt
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-aiohttp==1.0.5
ruff
bandit
pip-audit
mypy==1.10.0